mqtt_tele/requirements.txt
paho-mqtt==1.5.1
redis==3.5.3
requests==2.25.1
